Refactoring: user and dog forms

+++ kennel/dogs/views.py
@@ -64,7 +64,7 @@ def dogs_list(request):
 
 def dog_create(request):
     """Обрабатывает создание новой собаки через форму."""
-    form = DogForm(request.POST, request.FILES)
+    form = DogForm(request.POST or None, request.FILES or None)
     if request.method == "POST":
         if form.is_valid():
             form.save()
@@ -99,13 +99,17 @@ def dog_detail(request, pk: int):
 def dog_update(request, pk: int):
     """Обновляет данные у собаки."""
     dog_object = get_object_or_404(Dog, pk=pk)
-    file_path = dog_object.photo.path
+    if dog_object.photo:
+        file_path = dog_object.photo.path
+    else:
+        file_path = None
     if request.method == "POST":
         form = DogForm(request.POST, request.FILES, instance=dog_object)
         if form.is_valid():
             if "photo" in request.FILES:
-                if os.path.exists(file_path):
-                    os.remove(file_path)
+                if file_path:
+                    if os.path.exists(file_path):
+                        os.remove(file_path)
             dog_object = form.save(commit=False)
             dog_object.save()
             return redirect(reverse("dogs:dog_detail", args=[pk]))
